Refactor: move duplicated platform enums to shared KMP module
Move SongSortOrder, PracticeMode, PlayDirection, FretPosition, and
BeatType from Android-only packages to shared/data/ so both platforms
can use the same definitions without drift.

Fixes #400
