+++ app/src/test/java/com/baijum/ukufretboard/viewmodel/SongbookViewModelTest.kt
@@ -3,6 +3,7 @@ package com.baijum.ukufretboard.viewmodel
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
 import com.baijum.ukufretboard.data.ChordSheet
+import com.baijum.ukufretboard.data.SongSortOrder
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
