+++ shared/src/commonMain/kotlin/com/baijum/ukufretboard/data/BeatType.kt
@@ -1,0 +1,3 @@
+package com.baijum.ukufretboard.data
+
+enum class BeatType { ACCENT, NORMAL, MUTE }
