+++ shared/src/commonMain/kotlin/com/baijum/ukufretboard/data/SongSortOrder.kt
@@ -1,0 +1,8 @@
+package com.baijum.ukufretboard.data
+
+enum class SongSortOrder {
+    LAST_MODIFIED,
+    DATE_ADDED,
+    TITLE,
+    ARTIST,
+}
